blog-automation/requirements.txt
# Core HTTP client
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain for LLM
langchain>=0.1.0
langchain-openai>=0.0.5

# SerpAPI client
serpapi>=0.1.5

# Image processing (for future preview image generation)
Pillow>=10.2.0

# HTML parsing
beautifulsoup4>=4.12.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
